requirements.txt
torch --index-url https://download.pytorch.org/whl/cu117
torchvision --index-url https://download.pytorch.org/whl/cu117
scikit-learn
torchmetrics
numpy
pillow
tqdm
matplotlib
pandas
tensorboard
albumentations
opencv-python
